test: isolate CLI usage regression tests
